Moved portfolio class to singleton Began adding import of data from other portfolios

+++ branches/3.0.0/MyPersonalIndex/classes/assetAllocation.cpp
@@ -19,5 +19,8 @@ void assetAllocation::save(const int &portfolioID)
 
 void assetAllocation::remove() const
 {
+    if (this->id == -1)
+        return;
+
     queries::deleteItem(queries::table_AA, this->id);
 }
